setLiveEdgeSeekTarget has been moved in StreamProcessor class
